+++ src/main/kotlin/mod/item/skills/ArrowRain.kt
@@ -15,9 +15,9 @@ import net.minecraft.util.EnumHand
 import net.minecraft.world.World
 import kotlin.random.Random
 
-object ArrowRain : ItemSkill("arrowrain", 20, SkillRarity.UNCOMMON, false) {
-	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Float) {
-		if (StatusUtil().useMP(player, this.cost, savingRate)) {
+object ArrowRain : ItemSkill("arrowrain", 20.0, SkillRarity.UNCOMMON, false) {
+	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Double) {
+		if (StatusUtil.useMP(player, this.cost, savingRate)) {
 			if (!world.isRemote) {
 				val pos = player.rayTrace(15.0, 0.0F)?.blockPos!!
 				val itemstack = ItemStack(Items.ARROW)
