+++ src/main/kotlin/mod/item/skills/BlackHole.kt
@@ -9,9 +9,9 @@ import net.minecraft.util.EnumHand
 import net.minecraft.util.EnumParticleTypes
 import net.minecraft.world.World
 
-object BlackHole : ItemSkill("blackhole", 50, SkillRarity.RARE, false) {
-	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Float) {
-		if (StatusUtil().useMP(player, this.cost, savingRate)) {
+object BlackHole : ItemSkill("blackhole", 50.0, SkillRarity.RARE, false) {
+	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Double) {
+		if (StatusUtil.useMP(player, this.cost, savingRate)) {
 			val entityList = world.loadedEntityList
 			val livingList = arrayListOf<EntityLiving>()
 			val pos = player.rayTrace(15.0, 0.0F)!!.blockPos
