+++ src/main/kotlin/mod/item/skills/HealP.kt
@@ -7,9 +7,9 @@ import net.minecraft.entity.player.EntityPlayer
 import net.minecraft.util.EnumHand
 import net.minecraft.world.World
 
-object HealP : ItemSkill("heal+", 10, SkillRarity.UNCOMMON, false) {
-	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Float) {
-		if (StatusUtil().useMP(player, this.cost, savingRate)) {
+object HealP : ItemSkill("heal+", 10.0, SkillRarity.UNCOMMON, false) {
+	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Double) {
+		if (StatusUtil.useMP(player, this.cost, savingRate)) {
 			if (player.maxHealth <= player.health + 4) {
 				player.health = player.maxHealth
 			} else {
