+++ src/main/kotlin/mod/item/skills/HealPP.kt
@@ -7,9 +7,9 @@ import net.minecraft.entity.player.EntityPlayer
 import net.minecraft.util.EnumHand
 import net.minecraft.world.World
 
-object HealPP : ItemSkill("heal++", 25, SkillRarity.RARE, false) {
-	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Float) {
-		if (StatusUtil().useMP(player, this.cost, savingRate)) {
+object HealPP : ItemSkill("heal++", 25.0, SkillRarity.RARE, false) {
+	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Double) {
+		if (StatusUtil.useMP(player, this.cost, savingRate)) {
 			if (player.maxHealth <= player.health + 8) {
 				player.health = player.maxHealth
 			} else {
