+++ src/main/kotlin/mod/item/skills/Rage.kt
@@ -11,9 +11,9 @@ import net.minecraft.entity.player.EntityPlayer
 import net.minecraft.util.EnumHand
 import net.minecraft.world.World
 
-object Rage : ItemSkill("rage", 10, SkillRarity.UNCOMMON, false) {
-	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Float) {
-		if (StatusUtil().useMP(player, this.cost, savingRate)) {
+object Rage : ItemSkill("rage", 10.0, SkillRarity.UNCOMMON, false) {
+	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Double) {
+		if (StatusUtil.useMP(player, this.cost, savingRate)) {
 			GlobalScope.launch {
 				player.getEntityAttribute(SharedMonsterAttributes.ATTACK_DAMAGE).baseValue += 2
 				delay(10000)
