+++ src/main/kotlin/mod/item/skills/Shield.kt
@@ -11,9 +11,9 @@ import net.minecraft.entity.player.EntityPlayer
 import net.minecraft.util.EnumHand
 import net.minecraft.world.World
 
-object Shield : ItemSkill("shield", 15, SkillRarity.RARE, false) {
-	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Float) {
-		if (StatusUtil().useMP(player, this.cost, savingRate)) {
+object Shield : ItemSkill("shield", 15.0, SkillRarity.RARE, false) {
+	override fun skillFunction(world: World, player: EntityPlayer, handIn: EnumHand, savingRate: Double) {
+		if (StatusUtil.useMP(player, this.cost, savingRate)) {
 			GlobalScope.launch {
 				player.getEntityAttribute(SharedMonsterAttributes.ARMOR).baseValue += 2
 				delay(10000)
